requirements.txt
asgiref==3.7.2
Django==4.2.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
